int bug corrected

+++ app/recognizers/recognizer.py
@@ -57,7 +57,7 @@ class Recognizer():
         # resize the frame to have a width of 600 pixels (while
         # maintaining the aspect ratio), and then grab the image
         # dimensions
-        frame = imutils.resize(image, height=480 * RATIO_DETECTION)
+        frame = imutils.resize(image, height=int(480 * RATIO_DETECTION))
         (h, w) = frame.shape[:2]
 
         # construct a blob from the image
@@ -106,7 +106,10 @@ class Recognizer():
                 proba = preds[j]
                 name = le.classes_[j]
 
-                dicts.append({"box": (x1/RATIO_DETECTION, y1/RATIO_DETECTION, x2/RATIO_DETECTION, y2/RATIO_DETECTION),
+                dicts.append({"box": (int(x1/RATIO_DETECTION),
+                                      int(y1/RATIO_DETECTION),
+                                      int(x2/RATIO_DETECTION),
+                                      int(y2/RATIO_DETECTION)),
                               "confidence_face": confidence,
                               "name": name,
                               "confidence_name": proba})
